Add __delitem__
In line with how __getitem__ and __setitem__ work this cuts off the
branch "held" by the key, as well as removing the node data.

As discussed elsewhere the method naming is unusual, but if we intend to
change this it should be done elsewhere, after more thorough thinking
about the whole set of methods.